requirements.txt
Flask==2.0.1
Flask_SQLAlchemy==2.5.1
python-decouple==3.5
gunicorn==19.9.0